paging

+++ src/main/java/com/cos/blog/model/User.java
@@ -29,7 +29,7 @@ public class User {
 	
 	@Id//Primary key 선언
 	@GeneratedValue(strategy = GenerationType.IDENTITY)// jpa 넘버링전략이 아닌 프로젝트에서 연결된 DB의 넘버링 전략을 따라간다.
-	private Long id; //시퀀스 , auto_increment
+	private int id; //시퀀스 , auto_increment
 	
 	@Column(nullable = false, length = 30) // column의 조건
 	private String username;
